ALways return AdditiveColoringModel by ColorByColumnDialog
